taurus_engine/requirements.txt
boto==2.38.0
sqlalchemy==0.9.4
mysql-python==1.2.5
supervisor==3.1.3
uwsgi==2.0.12
alembic==0.6.7
web.py==0.37
msgpack-python==0.3.0
validictory==0.9.1
nta.utils
htmengine
paste==1.7.5.1

# required by tests
mock==1.0.1
